+++ src/refactoring/components/AdminPage.tsx
@@ -1,5 +1,5 @@
 import { Coupon, Product } from "../../types.ts";
-import { AddProductForm, CouponForm, Coupons, Products } from "./admin";
+import { AddCouponForm, AddProductForm, Coupons, Products } from "./admin";
 import { Layout } from "./shared";
 
 interface Props {
@@ -23,7 +23,7 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
         <div>
           <h2 className="text-2xl font-semibold mb-4">쿠폰 관리</h2>
           <div className="bg-white p-4 rounded shadow">
-            <CouponForm onSubmit={onCouponAdd} />
+            <AddCouponForm onSubmit={onCouponAdd} />
             <Coupons coupons={coupons} />
           </div>
         </div>
