+++ src/refactoring/components/admin/index.ts
@@ -1,7 +1,7 @@
+import AddCouponForm from "./AddCouponForm";
 import AddProductForm from "./AddProductForm";
-import CouponForm from "./CouponForm";
 import Coupons from "./Coupons";
 import EditProductForm from "./EditProductForm";
 import Products from "./Products";
 
-export { AddProductForm, CouponForm, Coupons, EditProductForm, Products };
+export { AddCouponForm, AddProductForm, Coupons, EditProductForm, Products };
